- Commented PXEngine a bit more - PXEngineGetViewWidth / PXEngineGetViewHeight now use floats since they are in POINTS (used to be in pixels) - Added ability to convert between PXTextureData <-> CGImage / UIImage - Added ability to take screenshot (dump of screen buffer) into a CGImage / UIImage - Small notes and chnages

+++ Pixelwave/Classes/Display/PXStage.m
@@ -202,8 +202,8 @@
 
 	//if (orient == orientation) return;
 
-	int viewWidth = PXEngineGetViewWidth();
-	int viewHeight = PXEngineGetViewHeight();
+	float viewWidth = PXEngineGetViewWidth();
+	float viewHeight = PXEngineGetViewHeight();
 
 	orientation = orient;
